requirements.txt
loguru==0.5.3
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
